refactor(dashboard): split the XvB/earnings/badges cluster out of views.py (#1105 Phase 3 D2, develop-v2)
D2 of the Phase 3 cut queue. The XvB, earnings and top-bar-badge builders move from
web/views.py into web/xvb_views.py, and their tests from tests/web/test_views.py into
tests/web/test_xvb_views.py. views.py 1919 -> 1129, test_views.py 3195 -> 1978.

The boundary is the forward closure of the cluster's dependency graph, not the cut map's
approximate line range: 27 top-level names in two contiguous runs with no non-member
interleaved. The only edge leaving the closure other than build_state is _LOW_HR_TITLE,
which build_hashrate also reads; views.py imports it back through the facade, the same
direction D1 established. xvb_views.py imports nothing from views.py.

Both moved runs are byte-identical to their source and the retained half of views.py is
byte-identical to base except the 16 imports the move made unused and the facade block.
The test move is 1:1 by pytest --collect-only, 2120 IDs before and after.

Moving build_badges broke a monkeypatch in tests/service/test_telegram_commands.py that
steered it through views.monero_is_local; that alias now binds the module which owns the
name. Every site reaching a moved or pruned name through the views module object was
enumerated and classified — four had to stay pointing at views, because they steer
build_state, which does not move.

The shared test-fixture builders are duplicated into the new module rather than shared:
pytest runs with --import-mode=importlib and the test tree has no __init__.py, so no
import route exists between two test modules. Declared in the new module's docstring;
consolidation tracked separately.
